reformat function files

+++ .config/omf/functions/kp.fish
@@ -1,14 +1,23 @@
-function kp -d "kill processes"
-  set -l __kp__pid (ps -ef | sed 1d | eval "fzf $FZF_DEFAULT_OPTS -m --header='[kill:process]'" | awk '{print $2}')
-  set -l __kp__kc $argv[1]
+# function to kill a process found with fzf
+function kp -d "kill processes found with fzf"
 
-  if test "x$__kp__pid" != "x"
-    if test "x$argv[1]" != "x"
-      echo $__kp__pid | xargs kill $argv[1]
-    else
-      echo $__kp__pid | xargs kill -9
+    # make sure fzf is installed
+    if ! type -q fzf
+        echo "error: fzf not installed"
+        return 1
     end
-    kp
-  end
+
+    # find process and kill it
+    set -l __kp__pid (ps -ef | sed 1d | eval "fzf $FZF_DEFAULT_OPTS -m --header='[kill:process]'" | awk '{print $2}')
+    set -l __kp__kc $argv[1]
+
+    if test "x$__kp__pid" != "x"
+        if test "x$argv[1]" != "x"
+            echo $__kp__pid | xargs kill $argv[1]
+        else
+            echo $__kp__pid | xargs kill -9
+        end
+        kp
+    end
 end
